requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
block-disposable-email==2.0.0
boto3==1.24.96
botocore==1.27.96
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.4.0
Django==3.2.15
django-appconf==1.0.5
django-cors-headers==3.13.0
django-cryptography==1.1
django-daraja==1.3.0
django-disposable-email-checker==2.0.3
django-extensions==3.2.1
django-filter==23.5
django-mpesa==2.0.14
django-rest-authemail==2.1.4
django-rest-knox==4.2.0
django-rest-logger @ git+https://github.com/Seedstars/django-rest-logger.git@86f99fed41eb993c580ab9143a13a228d2bd9e2d
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dynaconf==3.1.11
entrypoints==0.3
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
httplib2==0.14.0
hvac==1.0.2
idna==2.8
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
keyring==18.0.1
knox==0.1.14
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.6.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pbr==5.11.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
protobuf==3.6.1
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pycairo==1.21.0
pycparser==2.21
PyGObject==3.36.0
pyhcl==0.4.4
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-dateutil==2.7.3
python-decouple==3.8
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
s3transfer==0.6.0
SecretStorage==2.3.1
shortuuid==1.0.9
simplejson==3.16.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
systemd-python==234
testresources==2.0.1
uritemplate==4.1.1
urllib3==1.25.8
validators==0.20.0
virtualenv==20.16.5
wadllib==1.3.3
